✨ Feat : 점수 매기기 API

See merge request hyeseung2000/whereismyhome_final_backend!23

+++ src/main/java/com/newjeanssa/budongbudong/model/dto/review/ReviewAptDto.java
@@ -16,7 +16,7 @@ public class ReviewAptDto {
     private Long reviewId;
     private String aptId;
     private String reviewComment;
-    private int reviewScore;
+    private double reviewScore;
     private LocalDateTime createTime;
     private UserDto userDto;
 }
